Move random handler to random directory to make future scalability easier
