+++ src/lambdas/index.ts
@@ -1,2 +1,2 @@
-export { handler as randomGet } from './random-get'
+export { handler as randomGet } from './random/random-get'
 export { handler as breedsGet } from './breeds/breeds-get'
